serialization: Split into a bunch of smaller crates
The crates are now as follows:
* `serialization-core` re-exports basic types and traits from SCALE
* `serialization-tagged` adds support for tagged/direct encoding
* `serialization-tagged-derive` provides derive macros for the above
* `serialization` is a top-level crate re-exporting the above
