+++ serialization/core/Cargo.toml
@@ -1,0 +1,15 @@
+[package]
+name = "serialization-core"
+version = "0.1.0"
+edition = "2021"
+license = "MIT"
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+parity-scale-codec = {version = "3.1.2", features = ["derive", "chain-error"]}
+
+[dev-dependencies]
+arraytools = "0.1.5"
+hex-literal = "0.3.4"
+rand = "0.8.5"
