+++ serialization/core/src/lib.rs
@@ -17,7 +17,8 @@
 
 //! Blockchain data encoding and decoding tools
 
-// Re-export all the constituent parts
-pub use serialization_core::*;
-pub use serialization_tagged as tagged;
-pub use tagged::{DirectDecode, DirectEncode, Tag, Tagged};
+// Re-export SCALE traits
+pub use parity_scale_codec::{Codec, Decode, DecodeAll, Encode, Input, Output};
+
+// Re-export SCALE types
+pub use parity_scale_codec::{Compact, Error};
