+++ serialization/core/tests/simple_types.rs
@@ -1,6 +1,6 @@
 mod utils;
 
-use serialization::{Decode, Encode};
+use serialization_core::{Decode, Encode};
 use utils::{OptionWrapper, SimpleWrapper};
 
 #[test]
