+++ serialization/tagged/Cargo.toml
@@ -1,0 +1,19 @@
+[package]
+name = "serialization-tagged"
+version = "0.1.0"
+edition = "2021"
+license = "MIT"
+
+[dependencies]
+
+serialization-core = { path = '../core' }
+serialization-tagged-derive = { path = 'derive' }
+
+static_assertions = "1.1"
+
+[dev-dependencies]
+
+serialization = { path = '..' }
+
+parity-scale-codec = {version = "3.1.2", features = ["derive", "chain-error"]}
+proptest = "1.0"
