+++ serialization/tagged/derive/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "serialization-derive"
+name = "serialization-tagged-derive"
 version = "0.1.0"
 edition = "2021"
 license = "MIT"
